requirements.txt
torch
librosa
numpy
whisperx
opencc-python-reimplemented
PyYAML